3/6 캠프파이어 / 적 스포너
-> 낮에서 밤 / 적 패턴 다양 / 던전 퍼즐 / 적 추가 / 퀘스트 시스템 / 마법무기

+++ Assets/1_Script/Player/PlayerStatus.cs
@@ -89,7 +89,7 @@ public class PlayerStatus : MonoBehaviour, IDamageable
         }
     }
 
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         curHP += amount;
         curHP = Mathf.Clamp(curHP, 0, MaxHP);
